Extract lua methods
