+++ scripter/scripter.go
@@ -1,17 +1,12 @@
 package scripter
 
 import (
-	"fmt"
+	"bytes"
 	"github.com/BurntSushi/toml"
 	"github.com/honeytrap/honeytrap/abtester"
-	"github.com/honeytrap/honeytrap/utils/files"
+	"github.com/honeytrap/honeytrap/pushers"
 	"github.com/op/go-logging"
 	"net"
-	"time"
-	"encoding/json"
-	"github.com/honeytrap/honeytrap/pushers"
-	"github.com/honeytrap/honeytrap/event"
-	"bytes"
 )
 
 var (
@@ -100,89 +95,4 @@ func WithConfig(c toml.Primitive) ScripterFunc {
 	return func(scr Scripter) error {
 		return toml.PrimitiveDecode(c, scr)
 	}
-}
-
-//SetBasicMethods sets methods that can be called by each script, returning basic functionality
-func SetBasicMethods(s Scripter, c ScrConn, service string) {
-	c.SetStringFunction("getRemoteAddr", func() string { return c.GetConn().RemoteAddr().String() }, service)
-	c.SetStringFunction("getLocalAddr", func() string { return c.GetConn().LocalAddr().String() }, service)
-
-	c.SetStringFunction("getDatetime", func() string {
-		t := time.Now()
-		return fmt.Sprintf("%d-%02d-%02dT%02d:%02d:%02d-00:00\n",
-			t.Year(), t.Month(), t.Day(),
-			t.Hour(), t.Minute(), t.Second())
-	}, service)
-
-	c.SetStringFunction("getFileDownload", func() string {
-		params, _ := c.GetParameters([]string{"url", "path"}, service)
-
-		if err := files.Download(params["url"], params["path"]); err != nil {
-			log.Errorf("error downloading file: %s", err)
-			return "no"
-		}
-		return "yes"
-	}, service)
-
-	if ab, ok := c.(ScrAbTester); ok {
-		//In the script the function 'getAbTest(key)' can be called, returning a random result for the given key
-		c.SetStringFunction("getAbTest", func() string {
-			params, _ := c.GetParameters([]string{"key"}, service)
-
-			val, err := ab.GetAbTester().GetForGroup(service, params["key"], -1)
-			if err != nil {
-				return "_" //No response, _ so lua knows it has no ab-test
-			}
-
-			return val
-		}, service)
-	}
-
-	//In the script the function 'doLog(type, message)' can be called, with type = logging type and message the message
-	c.SetVoidFunction("doLog", func() {
-		params, _ := c.GetParameters([]string{"logType", "message"}, service)
-		logType := params["logType"]
-		message := params["message"]
-
-		if logType == "critical" {
-			log.Critical(message)
-		}
-		if logType == "debug" {
-			log.Debug(message)
-		}
-		if logType == "error" {
-			log.Error(message)
-		}
-		if logType == "fatal" {
-			log.Fatal(message)
-		}
-		if logType == "info" {
-			log.Info(message)
-		}
-		if logType == "notice" {
-			log.Notice(message)
-		}
-		if logType == "panic" {
-			log.Panic(message)
-		}
-		if logType == "warning" {
-			log.Warning(message)
-		}
-	}, service)
-
-	c.SetVoidFunction("channelSend", func() {
-		params, _ := c.GetParameters([]string{"data"}, service)
-		var data map[string]interface{}
-
-		json.Unmarshal([]byte(params["data"]), &data)
-
-		message := event.New()
-		for key, value := range data {
-			event.Custom(key, value)(message)
-		}
-		event.Custom("destination-ip", c.GetConn().LocalAddr().String())(message)
-		event.Custom("source-ip", c.GetConn().RemoteAddr().String())(message)
-
-		s.GetChannel().Send(message)
-	}, service)
 }
